Booking addons
Selecting a journey while booking will fetch that journeys addons, which then can be selected with checkboxes. after selectecting addons and making a booking, the booking will store these addons with their title(name) and price.

TODO:
Automatically select the first journey in the dropdown.
Update the price field by the total price of selected journy, number of people and selection of addons